# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Blickling Fieldbook API
servers:
  - url: /api
    description: Base API path
...
security:
  - cookieAuth: []
paths:
...
  /auth/setup:
    post:
      security: []
      operationId: completeSetup
      tags: [auth]
      summary: Complete the one-time secret-protected administrator setup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupInput"
      responses:
        "201":
          description: Setup completed and administrator signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
  /auth/login:
    post:
      security: []
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed

  # Users
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List observation categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (administrator/manager)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (administrator/manager)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  # Named Locations
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: blickling.sid
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          maxLength: 128
...
    SetupInput:
      type: object
      required: [setupSecret, name, email, password]
      properties:
        setupSecret:
          type: string
          minLength: 24
        name:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
        password:
          type: string
          minLength: 14
          maxLength: 128
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 14
          maxLength: 128
    AuthUser:
      type: object
      required: [id, name, email, role, propertyId, active, mustChangePassword]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [administrator, manager, team_member]
        propertyId:
          type: integer
        active:
          type: boolean
        mustChangePassword:
          type: boolean

    # Users
...
    Category:
      type: object
      required: [id, name, active, sortOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        displayColour:
          type: ["string", "null"]
        sortOrder:
          type: integer
        active:
          type: boolean
        propertyId:
          type: integer
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        displayColour:
          type: string
        sortOrder:
          type: integer
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        displayColour:
          type: ["string", "null"]
        sortOrder:
          type: integer
        active:
          type: boolean

    # Named Locations
...